even more refactoring

+++ src/data.zig
@@ -57,7 +57,7 @@ pub const InputData = struct {
 pub const erow = struct {
     chars: []u8,
     render: []u8,
-    highlight: []u16,
+    highlight: []u8,
 };
 
 //***** Values *****
